requirements.txt
opencv-python
paho-mqtt
ultralytics
SpeechRecognition
PyAudio
gTTS
requests
psutil
spotipy
piper-tts
